requirements.txt
django
psycopg2-binary
uwsgi
django-widget-tweaks
djangorestframework
drf-yasg
markdown
django-filter 
jsons
coverage
beautifulsoup4